requirements.txt
absl-py==2.1.0
androguard==3.4.0a1
appnope==0.1.4
asn1crypto==1.5.1
asttokens==3.0.0
black==25.1.0
click==8.1.8
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
debugpy==1.8.13
decorator==5.2.1
executing==2.2.0
filelock==3.17.0
fonttools==4.56.0
fsspec==2025.2.0
graphviz==0.20.3
grpcio==1.70.0
ipykernel==6.29.5
ipython==9.0.1
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
lxml==5.3.1
Markdown==3.7
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
mpmath==1.3.0
mypy-extensions==1.0.0
narwhals==1.30.0
nest-asyncio==1.6.0
networkx==3.4.2
numpy==2.2.3
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
packaging==24.2
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.6
plotly==6.0.0
polars==1.24.0
prettytable==3.15.1
prompt_toolkit==3.0.50
protobuf==6.30.0
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pydot==3.0.4
Pygments==2.19.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pyzmq==26.2.1
quark-engine==25.3.1
r2pipe==1.8.0
rzpipe==0.6.2
setuptools==75.8.2
six==1.17.0
stack-data==0.6.3
sympy==1.13.1
tensorboard==2.19.0
tensorboard-data-server==0.7.2
torch==2.6.0+cu126
torchaudio==2.6.0+cu126
torchvision==0.21.0+cu126
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
triton==3.2.0
typing_extensions==4.12.2
wcwidth==0.2.13
Werkzeug==3.1.3
